Add files via upload
In this code, I increased my scope by adding the functionality of multiple language support to it. It takes 2 inputs first is the Currency amount which has a limit up to 9 real numbers and a decimal precision up to 2 digits and the next is the integer value so that the language is to be chosen you need to  Enter 1 for English 2 for Hindi 3 for Marathi 4 for Bengali and 5 for Gujarati output respectively. In this code, the data structure with which I have worked is the  Hashmaps which is similar to a dictionary in python. It has a  Key and the associated value with it. For example, if the key is 1 the value associated with it is "one"  i.e. the textual output which is a string.